Moved window_main to separate folder
