+++ pytranscriber/control/ctr_main.py
@@ -13,14 +13,14 @@
 '''
 
 from PyQt5 import QtCore, QtWidgets
-from PyQt5.QtWidgets import QFileDialog, QMessageBox, QDialog, QActionGroup
+from PyQt5.QtWidgets import QFileDialog, QDialog, QActionGroup
 from PyQt5.QtCore import Qt
 from pathlib import Path
 from pytranscriber.model.param_autosub import Param_Autosub
 from pytranscriber.util.util import MyUtil
 from pytranscriber.control.thread_exec_autosub import Thread_Exec_Autosub
 from pytranscriber.control.thread_cancel_autosub import Thread_Cancel_Autosub
-from pytranscriber.gui.window_main import Ui_window
+from pytranscriber.gui.main.window_main import Ui_window
 from pytranscriber.control.ctr_db import CtrDB
 from pytranscriber.gui.message_util import MessageUtil
 import os
