+++ pytranscriber/gui/main/window_main.py
