+++ pytranscriber/gui/main/window_main.ui
